Fix unit tests

+++ src/test/java/com/fiap/fast_food_tc/unit/adapter/provider/OrdersDataProviderTest.java
@@ -102,6 +102,6 @@ class OrdersDataProviderTest {
 
         Short code = provider.getLastOrderCode();
 
-        assertEquals(0, code);
+        assertEquals((short)0, code);
     }
 }
